requirements.txt
google-cloud-bigquery>=3.10.0
pandas>=1.5
requests>=2.31
python-telegram-bot>=20.0
python-dotenv>=1.0
pytest>=7.0
